various:
* Multi user: its stateless, so it creates a uuid per file and all
  routes are per file with `default.comapeocat` as fallback
* Split get, post and watcher in separate files
